requirements.txt
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
elastic-transport==8.17.0
elasticsearch==8.17.1
Flask==3.1.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
numpy==2.2.3
opencv-python-headless==4.11.0.86
packaging==24.2
pillow==11.1.0
pytesseract==0.3.13
requests==2.32.3
urllib3==2.3.0
Werkzeug==3.1.3
